Fix potential memory corruption due to lifetime of default arguments
Functions which contain constant reference optional arguments must be
wary when assigning default values to them. Object construction in
headers here are bound to the scope of the function and will be
deallocated at the end. Our code suffered heavily from using this
pattern to facilite dependency injection that enabled testing.

For further reading see the following:
https://stackoverflow.com/questions/12554619/what-is-the-lifetime-of-a-default-argument-temporary-bound-to-a-reference-parame